feat(protos): add google.rpc protos
This commits adds proto files for some google.rpc types, which will
be used in our error package.

It also fixes a bug in the Rust protobuf generator rule where we were
passing the descriptor set files in the wrong way. We didn't catch it
before as it only occurs when multiple protos are built into a single
rust crate.

We also add a new disable_lint argument to pl_proto_library, that skips
generating the buf lint test. This is necessary as we will store some
external proto files and we don't want to change them.