feat(db): integrate FDB storage backend into main database type
This commit adds the integration with the FoundationDB backend created
in the previous commit into `pl_database::Db`.

We needed to change the range transaction API to make it work correctly
with the asynchronous nature of fdb. Now we have a for_each-style method,
needed as we can't return references to the byte slices returned from fdb.

We _could_ have implemented returning a mapped Stream instead, but it
would make the implementation more complex than what it already is, as
we would need to support sled also.

The fdb implementation of for_each_range pre-fetches the next part of the
range while processing each element from the current part. This means
that we process the range concurrently to the fetching. This was done
with the hypothesis that layers will consume most of the range, if not
all of it.

See documentation in code for more information.