feat(db/layers): add basic Collection layer
This commit adds a very basic "Collection" layer.

A `Collection<E>` is similar to `BTreeMap<Key, E>`, where `Key: TuplePack`.
The layer is responsible for providing an easy way of storing and
fetching entities by key or by a key prefix (i.e. a parent).

The implementation is so basic that we can't really use it in production
code, we still need a few thing:

1. A more controlled way of converting `ResourceName`s into the
underlying tree key. In the current implementation, each call
needs to convert the name into the correct tuple, which is very
error prone.

2. To be forward compatible with storage schema changes, we need
to add a initial byte of "metadata". This byte can be used to
provide information of what storage schema is used. Storage
schemas provide support for nice features as value compression and
encryption.
3. Memory pooling for encoding buffers.

These features will be added in the next commits.