feat(db/layers): add encoding metadata to Collection values
This commits adds the concept of "encoding metadata" into the value
stored in the database by the Collection layer.

This metadata can be used to provide information on how the value
was encoded, so that we can properly encode it. This is necessary
as we may add additional features into how we encode values in the
future, meaning that we need a way to differentiate between these
versions.

The first version is very basic, it simply states that the entity is
encoded using protobuf.