perf(db/layers): pre-allocate when possible in Collection::range
When we know the limit of entities to be returned by the range, we
can pre-allocate enough memory to store all of them. This commit
implements this optimization.

Note that more sophisticated pre-allocation logic can be implemented
on top of RangeOption, but in favor of keeping things simple here,
we leave these for future-us, our heros.